requirements.txt
pyyaml==6.0.1
geopandas==0.14.4
ipykernel==6.29.3
matplotlib==3.9.0